requirements.txt
nltk==3.8.1
pymilvus==2.5.4
sentence-transformers==3.4.1
tqdm==4.67.1
streamlit>=1.24.0
PyYAML==6.0.2
PyMuPDF==1.25.3
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
httpx>=0.24.0
groq>=0.4.0
